Move partition functions to the separate files.
