+++ os/kernel/src/filesystems/filesystem.c
@@ -1,49 +1,8 @@
 #include "filesystem.h"
 
-kvector partitions;
-
-bool filesystem_init()
-{
-    if(fdc_is_present() && floppy_is_inserted())
-    {
-        partition *floppy_partition = (partition*)heap_kernel_alloc(sizeof(partition), 0);
-        floppy_partition->type = filesystem_fat12;
-        floppy_partition->symbol = 'A';
-        floppy_partition->header = heap_kernel_alloc(512, 0);
-        
-        memcpy(floppy_partition->header, floppy_read_sector(0), 512);
-        kvector_add(&partitions, floppy_partition);
-        
-        fat_generic_set_current_partition(floppy_partition);
-        fat_init();
-    }
-}
-
-partition *filesystem_get_partition(char *path)
-{
-    char partition_symbol = path[0];
-    for (int i = 0; i < partitions.count; i++)
-    {
-        partition *partition_to_check = partitions.data[i];
-        if(partition_to_check->symbol == partition_symbol)
-        {
-            return partition_to_check;
-        }
-    }
-    
-    return 0;
-}
-
-void filesystem_remove_partition_from_path(char *old_path, char *new_path)
-{
-    int path_length = strlen(old_path);
-    memcpy(new_path, old_path + 2, path_length - 2);
-    new_path[path_length - 2] = 0;
-}
-
 bool filesystem_get_file_info(char *path, filesystem_file_info *file_info)
 {
-    partition *partition = filesystem_get_partition(path);
+    partition *partition = partitions_get(path);
     fat_generic_set_current_partition(partition);
     
     char path_without_partition[128];
@@ -54,7 +13,7 @@ bool filesystem_get_file_info(char *path, filesystem_file_info *file_info)
 
 bool filesystem_get_directory_info(char *path, filesystem_directory_info *directory_info)
 {
-    partition *partition = filesystem_get_partition(path);
+    partition *partition = partitions_get(path);
     fat_generic_set_current_partition(partition);
     
     char path_without_partition[128];
@@ -65,7 +24,7 @@ bool filesystem_get_directory_info(char *path, filesystem_directory_info *direct
 
 bool filesystem_read_file(char *path, uint8_t *buffer, uint32_t start_index, uint32_t length)
 {
-    partition *partition = filesystem_get_partition(path);
+    partition *partition = partitions_get(path);
     fat_generic_set_current_partition(partition);
     
     char path_without_partition[128];
@@ -76,7 +35,7 @@ bool filesystem_read_file(char *path, uint8_t *buffer, uint32_t start_index, uin
 
 uint32_t filesystem_get_entries_count_in_directory(char *path)
 {
-    partition *partition = filesystem_get_partition(path);
+    partition *partition = partitions_get(path);
     fat_generic_set_current_partition(partition);
     
     char path_without_partition[128];
@@ -87,7 +46,7 @@ uint32_t filesystem_get_entries_count_in_directory(char *path)
 
 bool filesystem_get_entries_in_directory(char *path, char **entries)
 {
-    partition *partition = filesystem_get_partition(path);
+    partition *partition = partitions_get(path);
     fat_generic_set_current_partition(partition);
     
     char path_without_partition[128];
@@ -98,7 +57,7 @@ bool filesystem_get_entries_in_directory(char *path, char **entries)
 
 bool filesystem_is_file(char *path)
 {
-    partition *partition = filesystem_get_partition(path);
+    partition *partition = partitions_get(path);
     fat_generic_set_current_partition(partition);
     
     char path_without_partition[128];
@@ -109,7 +68,7 @@ bool filesystem_is_file(char *path)
 
 bool filesystem_is_directory(char *path)
 {
-    partition *partition = filesystem_get_partition(path);
+    partition *partition = partitions_get(path);
     fat_generic_set_current_partition(partition);
     
     char path_without_partition[128];
@@ -130,7 +89,7 @@ uint32_t filesystem_get_total_space()
 
 bool filesystem_create_file(char *path)
 {
-    partition *partition = filesystem_get_partition(path);
+    partition *partition = partitions_get(path);
     fat_generic_set_current_partition(partition);
     
     char path_without_partition[128];
@@ -141,7 +100,7 @@ bool filesystem_create_file(char *path)
 
 bool filesystem_create_directory(char *path)
 {
-    partition *partition = filesystem_get_partition(path);
+    partition *partition = partitions_get(path);
     fat_generic_set_current_partition(partition);
     
     char path_without_partition[128];
@@ -152,7 +111,7 @@ bool filesystem_create_directory(char *path)
 
 bool filesystem_delete_file(char *path)
 {
-    partition *partition = filesystem_get_partition(path);
+    partition *partition = partitions_get(path);
     fat_generic_set_current_partition(partition);
     
     char path_without_partition[128];
@@ -163,7 +122,7 @@ bool filesystem_delete_file(char *path)
 
 bool filesystem_delete_directory(char *path)
 {
-    partition *partition = filesystem_get_partition(path);
+    partition *partition = partitions_get(path);
     fat_generic_set_current_partition(partition);
     
     char path_without_partition[128];
@@ -174,7 +133,7 @@ bool filesystem_delete_directory(char *path)
 
 bool filesystem_rename_file(char *path, char *new_name)
 {
-    partition *partition = filesystem_get_partition(path);
+    partition *partition = partitions_get(path);
     fat_generic_set_current_partition(partition);
     
     char path_without_partition[128];
@@ -185,7 +144,7 @@ bool filesystem_rename_file(char *path, char *new_name)
 
 bool filesystem_rename_directory(char *path, char *new_name)
 {
-    partition *partition = filesystem_get_partition(path);
+    partition *partition = partitions_get(path);
     fat_generic_set_current_partition(partition);
     
     char path_without_partition[128];
@@ -196,7 +155,7 @@ bool filesystem_rename_directory(char *path, char *new_name)
 
 bool filesystem_save_to_file(char *path, char *buffer, int size)
 {
-    partition *partition = filesystem_get_partition(path);
+    partition *partition = partitions_get(path);
     fat_generic_set_current_partition(partition);
     
     char path_without_partition[128];
@@ -207,11 +166,18 @@ bool filesystem_save_to_file(char *path, char *buffer, int size)
 
 bool filesystem_append_to_file(char *path, char *buffer, int size)
 {
-    partition *partition = filesystem_get_partition(path);
+    partition *partition = partitions_get(path);
     fat_generic_set_current_partition(partition);
     
     char path_without_partition[128];
     filesystem_remove_partition_from_path(path, path_without_partition);
     
     return fat_generic_append_to_file(path_without_partition, buffer, size);
+}
+
+void filesystem_remove_partition_from_path(char *old_path, char *new_path)
+{
+    int path_length = strlen(old_path);
+    memcpy(new_path, old_path + 2, path_length - 2);
+    new_path[path_length - 2] = 0;
 }
