+++ os/kernel/src/filesystems/filesystem.h
@@ -8,14 +8,10 @@
 #include "klibrary/kvector.h"
 #include "drivers/floppy/floppy.h"
 #include "drivers/harddisk/harddisk.h"
-#include "partition.h"
 
 // This is the common entry point for all filesystems. Now it will support only FAT, so every
 // method with fat_ prefix will be called directly. It will be changed in the future if there
 // will be more supported file systems.
-
-bool filesystem_init();
-partition *filesystem_get_partition(char *path);
 void filesystem_remove_partition_from_path(char *old_path, char *new_path);
 bool filesystem_get_file_info(char *path, filesystem_file_info *file_info);
 bool filesystem_get_directory_info(char *path, filesystem_directory_info *directory_info);
@@ -35,5 +31,7 @@ bool filesystem_rename_file(char *path, char *new_name);
 bool filesystem_rename_directory(char *path, char *new_name);
 bool filesystem_save_to_file(char *path, char *buffer, int size);
 bool filesystem_append_to_file(char *path, char *buffer, int size);
+
+void filesystem_remove_partition_from_path(char *old_path, char *new_path);
 
 #endif
