+++ os/kernel/src/filesystems/partitions/device_type.h
@@ -1,0 +1,11 @@
+#ifndef DEVICE_TYPE_H
+#define DEVICE_TYPE_H
+
+typedef enum device_type
+{
+    device_type_unknown,
+    device_type_floppy,
+    device_type_harddisk
+} device_type;
+
+#endif
